Refactor video service into a modular FileService:
- Extract file-related operations from `VideoService` to a new `FileService` for better separation of concerns.
- Rename `backend/src/services/video.py` to `backend/src/services/files.py` and update imports accordingly.
- Adjust API logic to utilize `FileService`, including video upload, retrieval, and deletion endpoints.
- Improve code organization with clearer abstraction and modularization of file handling logic.
